+++ random_profile/main.py
@@ -8,12 +8,15 @@ github : codeperfectplus
 import os
 import uuid
 import random
+from typing import List, Tuple
 
 from random_profile.utils import ipv4_gen
 from random_profile.utils import load_txt_file
 from random_profile.utils import generate_dob_age
 from random_profile.utils import generate_random_height_weight
 from random_profile.utils import ASSETS_DIR
+
+VERSION = "1.0.1"
 
 fname_txt = os.path.join(ASSETS_DIR, "fnames.txt")
 lname_txt = os.path.join(ASSETS_DIR, "lnames.txt")
@@ -23,6 +26,7 @@ street_names_txt = os.path.join(ASSETS_DIR, "street_names.txt")
 cities_name_txt = os.path.join(ASSETS_DIR, "cities_name.txt")
 states_names_txt = os.path.join(ASSETS_DIR, "states_names.txt")
 job_titles_txt = os.path.join(ASSETS_DIR, "job_titles.txt")
+states_hash_json = os.path.join(ASSETS_DIR, "states_hash.json")
 
 # loading data from txt files
 fname = load_txt_file(fname_txt)
@@ -36,84 +40,138 @@ job_titles = load_txt_file(job_titles_txt)
 
 
 class RandomProfile(object):
+    """ Random Profile Generator """
     def __init__(self, num: int = 1):
-        '''
-        num = Total No. of Name You Want To Print
-        default is 1
-        To Print More Than one Name Change value of num
-        '''
         self.num = num
 
-    def first_name(self, num: int = None) -> list:
-        if num is None:
-            num = self.num
-        first_name_list = [random.choice(fname) for _ in range(num)]
-        return first_name_list
+    def __str__(self) -> str:
+        return f"Random Profile Generator version {VERSION}"
+
+    def __repr__(self) -> str:
+        return f"RandomProfile(num={self.num})"
+
+    def __call__(self, num: int = None) -> List[dict]:
+        return self.full_profile(num)
+
+    def __iter__(self):
+        yield self.full_profile()
+
+    def __next__(self):
+        yield self.full_profile()
+
+    def __len__(self):
+        return self.num
+
+    def __getitem__(self, index):
+        return self.full_profile()[index]
+
+    def first_name(self, num: int = None) -> List[str]:
+        num = self.num if num is None else num
+        if num == 1 or num is None:
+            return random.choice(fname)
+        return random.choices(fname, k=num)
 
     def last_name(self, num: int = None) -> list:
-        if num is None:
-            num = self.num
-        last_name_list = [random.choice(lname) for _ in range(num)]
-        return last_name_list
+        num = self.num if num is None else num
+        if num == 1 or num is None:
+            return random.choice(lname)
+        return random.choices(lname, k=num)
 
-    def full_name(self, num: int = None) -> list:
-        if num is None:
-            num = self.num
-        full_name_list = [random.choice(
-            fname) + ' ' + random.choice(lname) for _ in range(num)]
-        return full_name_list
+    def full_name(self, num: int = None) -> List[str]:
+        num = self.num if num is None else num
+        if num == 1 or num is None:
+            return f"{random.choice(fname)} {random.choice(lname)}"
+        return [random.choice(fname) + ' ' + random.choice(lname) for _ in range(num)]
 
-    def full_profile(self, num: int = None) -> list:
-        if num is None:
-            num = self.num
-        profile_list = []
+    def ip_address(self, num: int = None) -> List[str]:
+        num = self.num if num is None else num
+        if num == 1 or num is None:
+            return ipv4_gen()
+        return [ipv4_gen() for _ in range(num)]
+
+    def job_title(self, num: int = None) -> List[str]:
+        num = self.num if num is None else num
+        if num == 1 or num is None:
+            return random.choice(job_titles)
+        return random.choices(job_titles, k=num)
+
+    def blood_type(self, num: int = None) -> List[str]:
+        num = self.num if num is None else num
+        if num == 1 or num is None:
+            return random.choice(blood_types)
+        return random.choices(blood_types, k=num)
+
+    def hair_color(self, num: int = None) -> List[str]:
+        num = self.num if num is None else num
+        if num == 1 or num is None:
+            return random.choice(hair_colors)
+        return random.choices(hair_colors, k=num)
+
+    def dob_age(self, num: int = None) -> List[Tuple[str, int]]:
+        num = self.num if num is None else num
+        if num == 1 or num is None:
+            return generate_dob_age()
+        return [generate_dob_age() for _ in range(num)]
+
+    def height_weight(self, num: int = None) -> List[Tuple[int, int]]:
+        num = self.num if num is None else num
+        if num == 1 or num is None:
+            return generate_random_height_weight()
+        return [generate_random_height_weight() for _ in range(num)]
+
+    def generate_address(self, num: int = None) -> List[str]:
+        num = self.num if num is None else num
+        address_list = []
         for _ in range(num):
-            unique_id = uuid.uuid4().hex
-            first = random.choice(fname)
-            last = random.choice(lname)
-            hair_color = random.choice(hair_colors)
-            blood_type = random.choice(blood_types)
-            full_name = first + ' ' + last
-            phone = f'+1-{random.randint(300, 500)}-{random.randint(800, 999)}-{random.randint(1000,9999)}'
-            job_title = random.choice(job_titles)
-            ip_address = ipv4_gen()
-
-            dob, age = generate_dob_age()
-            height, weight = generate_random_height_weight()
-
             street_num = random.randint(100, 999)
             street = random.choice(street_names)
             city = random.choice(cities_name)
             state = random.choice(states_names)
             zip_code = random.randint(10000, 99999)
 
-            address = f'{street_num} {street} St. {city} {state} {zip_code}'
-            email = first.lower() + last.lower() + '@example.com'
+            address = {
+                "street_num": street_num,
+                "street": street,
+                "city": city,
+                "state": state,
+                "zip_code": zip_code
+            }
+            address_list.append(address)
 
-            profile_dict = {}
-            profile_dict['id'] = unique_id
-            profile_dict['first_name'] = first
-            profile_dict['last_name'] = last
-            profile_dict['hair_color'] = hair_color
-            profile_dict['blood_type'] = blood_type
-            profile_dict['full_name'] = full_name
-            profile_dict['DOB'] = dob
-            profile_dict['age'] = age
-            profile_dict['height'] = height
-            profile_dict['weight'] = weight
-            profile_dict['phone'] = phone
-            profile_dict['address'] = address
-            profile_dict['email'] = email
-            profile_dict['job_title'] = job_title
-            profile_dict['ip_address'] = ip_address
-            profile_list.append(profile_dict)
+        return address_list
+
+    def full_profile(self, num: int = None) -> List[dict]:
+        num = self.num if num is None else num
+        profile_list = []
+        for _ in range(num):
+            phone_number = f'+1-{random.randint(300, 500)}-{random.randint(800, 999)}-{random.randint(1000,9999)}'
+            dob, age = self.dob_age(num=1)
+            height, weight = self.height_weight(num=1)
+            address = self.generate_address()[0]
+            full_address = '{} {} {}, {} {}'.format(address['street_num'],
+                                                    address['street'],
+                                                    address['city'],
+                                                    address['state'],
+                                                    address['zip_code'])
+
+            profile = {}
+            profile['id'] = str(uuid.uuid4())
+            profile['first_name'] = self.first_name(num=1)
+            profile['last_name'] = self.first_name(num=1)
+            profile['full_name'] = profile['first_name'] + ' ' + profile['last_name']
+            profile['job_title'] = self.job_title(num=1)
+            profile['dob'] = dob
+            profile['age'] = age
+            profile['phone_number'] = phone_number
+            profile['email'] = profile['first_name'].lower() + profile['last_name'].lower() + '@example.com'
+            profile['address'] = address
+            profile['full_address'] = full_address
+            profile['blood_type'] = self.blood_type(num=1)
+            profile['height'] = height
+            profile['weight'] = weight
+            profile['hair_color'] = self.hair_color(num=1)
+            profile['ip_address'] = self.ip_address(num=1)
+
+            profile_list.append(profile)
 
         return profile_list
-
-    def ipv4(self) -> list:
-        ip_list = [ipv4_gen() for _ in range(self.num)]
-        return ip_list
-
-    def job_title(self) -> list:
-        job_title_list = [random.choice(job_titles) for _ in range(self.num)]
-        return job_title_list